requirements.txt
tensorflow==2.13.0
numpy==1.24.3
opencv-python==4.8.0
pillow==10.0.0
